requirements.txt
feedparser>=6.0
requests>=2.28
pyyaml>=6.0
# Optional — required for LLM-driven bespoke UC generation. Two paths:
#   ANTHROPIC_API_KEY=...               → uses anthropic SDK (pay per token)
#   USECASEINTEL_USE_CLAUDE_OAUTH=1     → uses claude-agent-sdk + the user's
#                                         Claude Code session (Pro / Max
#                                         subscription, no separate billing)
# Pipeline runs without either; LLM UCs just don't get generated.
anthropic>=0.30; python_version >= "3.9"
claude-agent-sdk>=0.1; python_version >= "3.10"

# Sigma rule tooling. Core pysigma parses + validates rules; backends are
# optional and only required if you want to compile rules to a target
# platform (cheat-sheet "compile to..." button uses these). Each backend
# is its own pip package.
pysigma>=0.10
pysigma-backend-microsoft365defender>=0.1; python_version >= "3.9"
pysigma-backend-splunk>=1.0; python_version >= "3.9"
